Repackaged Request implicit class to ParamsAccessor trait
Ressource now use directly InMessage instead of former Request class
